+++ APITX/app/vistas/vistaUsuario.py
@@ -12,28 +12,32 @@ from django.contrib.auth.hashers import PBKDF2PasswordHasher
 from django.core.exceptions import ObjectDoesNotExist
 from app import permisos
 from app.vistas import autentificacion
+from rest_framework.authtoken.models import Token
 
 @api_view(['GET'])
 @authentication_classes((SessionAuthentication, BasicAuthentication,))
 @permission_classes((IsAuthenticated,))
-def autenticar(request, tk, format=None):
+def autenticar(request, format=None):
 
     """
     Este metodo devuelve los datos de un usuario que se esta logeando
     """
-    usuario = autentificacion.autenticacion(tk)
-    if usuario.has_perms(permisos.duenios):
-        if request.method == 'GET':
-            try:
-                objetoUsuario = User.objects.get(username=request.user)
-            except ObjectDoesNotExist:
-                return Response(status=status.HTTP_404_NOT_FOUND)
+    try:
+        objToken = Token.objects.get(key=request.query_params.get('tk'))
+        usuario = User.objects.get(pk=objToken.user.id)
+    except ObjectDoesNotExist:
+        content = {'Datos incorrectos': 'El token enviado no coincide para ningun usuario'}
+        return Response(content, status=status.HTTP_403_FORBIDDEN)
 
-            duenio = TxdDuenio.objects.get(usuario=objetoUsuario.id)
-            serializador = TxdDuenioS(duenio)
-            return Response(serializador.data)
-    else:
-        return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
+    if request.method == 'GET':
+        try:
+            objetoUsuario = User.objects.get(username=request.user)
+        except ObjectDoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        duenio = TxdDuenio.objects.get(usuario=objetoUsuario.id)
+        serializador = TxdDuenioS(duenio)
+        return Response(serializador.data)
 
 
 
@@ -48,13 +52,19 @@ def autenticar(request, tk, format=None):
 
 
 @api_view(['POST'])
-def crear_usuario(request, tk):
+def crear_usuario(request):
     """
     este metodo crea un nuevo usuario y retorna los datos creados
     """
-    usuario = autentificacion.autenticacion(tk)
-    if usuario.has_perms(permisos.duenios):
-        if request.method == 'POST':
+    try:
+        objToken = Token.objects.get(key=request.query_params.get('tk'))
+        usuario = User.objects.get(pk=objToken.user.id)
+    except ObjectDoesNotExist:
+        content = {'Datos incorrectos': 'El token enviado no coincide para ningun usuario'}
+        return Response(content, status=status.HTTP_403_FORBIDDEN)
+
+    if request.method == 'POST':
+        if usuario.has_perm('app.add_user'):
             serializador = UserSerializer(data = request.data)
             if serializador.is_valid():
                 user = User.objects.create_user(username=request.data.get('username'),
@@ -76,28 +86,38 @@ def crear_usuario(request, tk):
                 serializador = UserSerializer(objetoUsuario)
                 return Response(serializador.data, status=status.HTTP_201_CREATED)
             else:
-                return Response(status=status.HTTP_400_BAD_REQUEST)
-        return Response(serializador.errors, status=status.HTTP_400_BAD_REQUEST)
-    else:
-        return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
+                return Response(serializador.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            content = {'Permiso denegado': 'El usuario no tiene permisos para ingresar datos'}
+            return Response(content, status=status.HTTP_403_FORBIDDEN)
+
+    return Response(status=status.HTTP_400_BAD_REQUEST)
+
+
 
 @api_view(['GET', 'PUT','DELETE'])
-def detalle_usuario(request, pk, tk):
+def detalle_usuario(request, pk):
         """
         Actualiza, elimina un objeto segun su id
         """
-        usuario = autentificacion.autenticacion(tk)
-        if usuario.has_perms(permisos.duenios):
-            try:
-                obUsuario =  User.objects.get(pk = pk)
-            except ObjectDoesNotExist:
-                return Response(status=status.HTTP_404_NOT_FOUND)
+        try:
+            objToken = Token.objects.get(key=request.query_params.get('tk'))
+            usuario = User.objects.get(pk=objToken.user.id)
+        except ObjectDoesNotExist:
+            content = {'Datos incorrectos': 'El token enviado no coincide para ningun usuario'}
+            return Response(content, status=status.HTTP_403_FORBIDDEN)
 
-            if request.method == 'GET':
-                serUsuario = UserSerializer(obUsuario)
-                return Response(serUsuario.data)
+        try:
+            obUsuario =  User.objects.get(pk = pk)
+        except ObjectDoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
 
-            elif request.method == 'PUT':
+        if request.method == 'GET':
+            serUsuario = UserSerializer(obUsuario)
+            return Response(serUsuario.data)
+
+        elif request.method == 'PUT':
+            if usuario.has_perm('app.change_user'):
                 serUsuario = UserSerializer(obUsuario, data=request.data)
                 if serUsuario.is_valid():
                     serUsuario.save()
@@ -106,29 +126,44 @@ def detalle_usuario(request, pk, tk):
                     user.save()
                     return Response(serUsuario.data)
                 return Response(serUsuario.errors,status=status.HTTP_400_BAD_REQUEST)
+            else:
+                content = {'Permiso denegado': 'El usuario no tiene permisos para editar datos'}
+                return Response(content, status=status.HTTP_403_FORBIDDEN)
 
-            elif request.method == 'DELETE':
+
+        elif request.method == 'DELETE':
+            if usuario.has_perm('app.delete_user'):
                 objeto.delete()
                 return Response(status=status.HTTP_204_NO_CONTENT)
-        else:
-            return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
+            else:
+                content = {'Permiso denegado': 'El usuario no tiene permisos para eliminar datos'}
+                return Response(content, status=status.HTTP_403_FORBIDDEN)
+
+
 
 @api_view(['GET','POST'])
-def lista_usuario(request,tk):
+def lista_usuario(request):
         """
         Actualiza, elimina un objeto segun su id
         """
-        usuario = autentificacion.autenticacion(tk)
-        if usuario.has_perms(permisos.duenios):
-            try:
-                obUsuario =  User.objects.all()
-            except ObjectDoesNotExist:
-                return Response(status=status.HTTP_404_NOT_FOUND)
+        try:
+            objToken = Token.objects.get(key=request.query_params.get('tk'))
+            usuario = User.objects.get(pk=objToken.user.id)
+        except ObjectDoesNotExist:
+            content = {'Datos incorrectos': 'El token enviado no coincide para ningun usuario'}
+            return Response(content, status=status.HTTP_403_FORBIDDEN)
 
-            if request.method == 'GET':
-                serUsuario = UserSerializer(obUsuario,many=True)
-                return Response(serUsuario.data)
-            elif request.method == 'POST':
+        try:
+            obUsuario =  User.objects.all()
+        except ObjectDoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if request.method == 'GET':
+            serUsuario = UserSerializer(obUsuario,many=True)
+            return Response(serUsuario.data)
+
+        elif request.method == 'POST':
+            if usuario.has_perm('app.add_user'):
                 serializador = UserSerializer(data = request.data)
                 if serializador.is_valid():
                     user = User.objects.create_user(username=request.data.get('username'),
@@ -145,43 +180,54 @@ def lista_usuario(request,tk):
                     return Response(serializador.data, status=status.HTTP_201_CREATED)
                 else:
                     return Response(serializador.errors,status=status.HTTP_400_BAD_REQUEST)
+            else:
+                content = {'Permiso denegado': 'El usuario no tiene permisos para ingresar datos'}
+                return Response(content, status=status.HTTP_403_FORBIDDEN)
 
-        else:
-            return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
+
 
 @api_view(['GET'])
-def Usuarios_Group(request, pk, tk):
+def Usuarios_Group(request, pk):
         """
         Actualiza, elimina un objeto segun su id
         """
-        usuario = autentificacion.autenticacion(tk)
-        if usuario.has_perms(permisos.duenios):
-            try:
-                usuarios = User.objects.filter(groups=pk)
-            except ObjectDoesNotExist:
-                return Response(status=status.HTTP_404_NOT_FOUND)
+        try:
+            objToken = Token.objects.get(key=request.query_params.get('tk'))
+            usuario = User.objects.get(pk=objToken.user.id)
+        except ObjectDoesNotExist:
+            content = {'Datos incorrectos': 'El token enviado no coincide para ningun usuario'}
+            return Response(content, status=status.HTTP_403_FORBIDDEN)
 
-            if request.method == 'GET':
-                serializador = UserSerializer(usuarios, many=True)
-                return Response(serializador.data)
-            else:
-                return Response(serUsuario.errors,status=status.HTTP_400_BAD_REQUEST)
+        try:
+            usuarios = User.objects.filter(groups=pk)
+        except ObjectDoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if request.method == 'GET':
+            serializador = UserSerializer(usuarios, many=True)
+            return Response(serializador.data)
         else:
-            return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
+            return Response(serUsuario.errors,status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET', 'PUT','DELETE'])
-def CambiarEstado(request, pk, var, tk):
+def CambiarEstado(request, pk, var):
         """
         Actualiza, elimina un objeto segun su id
         """
-        usuario = autentificacion.autenticacion(tk)
-        if usuario.has_perms(permisos.duenios):
-            try:
-                obUsuario =  User.objects.get(pk=pk)
-            except ObjectDoesNotExist:
-                return Response(status=status.HTTP_404_NOT_FOUND)
+        try:
+            objToken = Token.objects.get(key=request.query_params.get('tk'))
+            usuario = User.objects.get(pk=objToken.user.id)
+        except ObjectDoesNotExist:
+            content = {'Datos incorrectos': 'El token enviado no coincide para ningun usuario'}
+            return Response(content, status=status.HTTP_403_FORBIDDEN)
 
-            if request.method == 'PUT':
+        try:
+            obUsuario =  User.objects.get(pk=pk)
+        except ObjectDoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if request.method == 'PUT':
+            if usuario.has_perms('app.change_user'):
                 if(var==0):
                     U = UserSerializer(obUsuario)
                     dato = U.data
@@ -205,23 +251,30 @@ def CambiarEstado(request, pk, var, tk):
                 else:
                     return Response(status=status.HTTP_400_BAD_REQUEST)
             else:
-                return Response(status=status.HTTP_400_BAD_REQUEST)
-        else:
-            return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
+                content = {'Permiso denegado': 'El usuario no tiene permisos para editar datos'}
+                return Response(content, status=status.HTTP_403_FORBIDDEN)
+
+
 
 @api_view(['GET', 'PUT','DELETE'])
-def cambiarGrupo(request, pk, tk):
+def cambiarGrupo(request, pk):
         """
         Actualiza, elimina un objeto segun su id
         """
-        usuario = autentificacion.autenticacion(tk)
-        if usuario.has_perms(permisos.duenios):
-            try:
-                obUsuario =  User.objects.get(pk = pk)
-            except ObjectDoesNotExist:
-                return Response(status=status.HTTP_404_NOT_FOUND)
+        try:
+            objToken = Token.objects.get(key=request.query_params.get('tk'))
+            usuario = User.objects.get(pk=objToken.user.id)
+        except ObjectDoesNotExist:
+            content = {'Datos incorrectos': 'El token enviado no coincide para ningun usuario'}
+            return Response(content, status=status.HTTP_403_FORBIDDEN)
 
-            if request.method == 'PUT':
+        try:
+            obUsuario =  User.objects.get(pk = pk)
+        except ObjectDoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if request.method == 'PUT':
+            if usuario.has_perm('app.change_user'):
                 if "idGrupoNuevo" in request.data:
                     serUsuario = UserSerializer(obUsuario)
                     user=serUsuario.data
@@ -243,6 +296,8 @@ def cambiarGrupo(request, pk, tk):
                 else:
                     return Response(status=status.HTTP_400_BAD_REQUEST)
             else:
-                return Response(status=status.HTTP_400_BAD_REQUEST)
+                content = {'Permiso denegado': 'El usuario no tiene permisos para editar datos'}
+                return Response(content, status=status.HTTP_403_FORBIDDEN)
+
         else:
-            return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
+            return Response(status=status.HTTP_400_BAD_REQUEST)
